+++ src/notification/module.ts
@@ -6,11 +6,12 @@ import NotificationService from "src/notification/application/service";
 
 import NotificationORMEntity from "src/notification/infrastructure/ormEntity";
 
-import NotificationController from "src/notification/presentation/controller";
+import NotificationHttpController from "./presentation/controllers/httpController";
+import NotificationsMessageController from "./presentation/controllers/messageController";
 
 @Module({
     providers: [NotificationService],
-    controllers: [NotificationController],
+    controllers: [NotificationHttpController, NotificationsMessageController],
     imports: [
         // 환경 변수 글로벌 설정
         ConfigModule.forRoot(),
